Keep serialization_fwd lean
