+++ hpx/runtime/serialization/serialization_fwd.hpp
@@ -10,38 +10,14 @@
 #include <hpx/config.hpp>
 #include <hpx/util/detail/pp_strip_parens.hpp>
 
-#if defined(HPX_INTEL_VERSION) && (HPX_INTEL_VERSION < 1400)
-#include <boost/shared_ptr.hpp>
-#else
-#include <memory>
-#endif
-
-#include <boost/type_traits/add_const.hpp>
-#include <boost/cstdint.hpp>
+#include <type_traits>
 
 namespace hpx { namespace serialization
 {
-    namespace detail
-    {
-        struct ptr_helper;
-#if defined(HPX_INTEL_VERSION) && (HPX_INTEL_VERSION < 1400)
-        typedef boost::shared_ptr<ptr_helper> ptr_helper_ptr;
-#else
-        typedef std::unique_ptr<ptr_helper> ptr_helper_ptr;
-#endif
-    }
-
     class access;
     struct input_archive;
     struct output_archive;
     struct binary_filter;
-
-    HPX_FORCEINLINE
-    void register_pointer(input_archive & ar, boost::uint64_t pos,
-        detail::ptr_helper_ptr helper);
-
-    template <typename Helper>
-    Helper & tracked_pointer(input_archive & ar, boost::uint64_t pos);
 
     template <typename T>
     output_archive & operator<<(output_archive & ar, T const & t);
@@ -67,32 +43,34 @@ namespace hpx { namespace serialization
         save(ar, 0);                                                                \
     }                                                                               \
 /**/
+
 #define HPX_SERIALIZATION_SPLIT_FREE(T)                                             \
-    HPX_FORCEINLINE                                                               \
+    HPX_FORCEINLINE                                                                 \
     void serialize(hpx::serialization::input_archive & ar, T & t, unsigned)         \
     {                                                                               \
         load(ar, t, 0);                                                             \
     }                                                                               \
-    HPX_FORCEINLINE                                                               \
+    HPX_FORCEINLINE                                                                 \
     void serialize(hpx::serialization::output_archive & ar, T & t, unsigned)        \
     {                                                                               \
-        save(ar, const_cast<boost::add_const<T>::type &>(t), 0);                    \
+        save(ar, const_cast<std::add_const<T>::type &>(t), 0);                      \
     }                                                                               \
 /**/
+
 #define HPX_SERIALIZATION_SPLIT_FREE_TEMPLATE(TEMPLATE, ARGS)                       \
     HPX_UTIL_STRIP(TEMPLATE)                                                        \
-    HPX_FORCEINLINE                                                               \
+    HPX_FORCEINLINE                                                                 \
     void serialize(hpx::serialization::input_archive & ar,                          \
             HPX_UTIL_STRIP(ARGS) & t, unsigned)                                     \
     {                                                                               \
         load(ar, t, 0);                                                             \
     }                                                                               \
     HPX_UTIL_STRIP(TEMPLATE)                                                        \
-    HPX_FORCEINLINE                                                               \
+    HPX_FORCEINLINE                                                                 \
     void serialize(hpx::serialization::output_archive & ar,                         \
             HPX_UTIL_STRIP(ARGS) & t, unsigned)                                     \
     {                                                                               \
-        save(ar, const_cast<typename boost::add_const                               \
+        save(ar, const_cast<typename std::add_const                                 \
                 <HPX_UTIL_STRIP(ARGS)>::type &>(t), 0);                             \
     }                                                                               \
 /**/
